+++ docs/src/whats-new.md
@@ -7,6 +7,27 @@ description: Learn about the latest features and improvements in LogLayer
 
 - [`loglayer` Changelog](/core-changelogs/loglayer-changelog)
 
+## May 10, 2025
+
+*It is advised that if you update `loglayer`, you should update any plugins and transports to the latest versions as well.*
+
+- Added the following new methods to control log levels in `loglayer`:
+  - `setLogLevel()`
+  - `enableIndividualLevel()`
+  - `disableIndividualLevel()`
+  - `isLogLevelEnabled()`
+
+See [the documentation](/logging-api/basic-logging#enabling-disabling-logging) for more information.
+
+- Fixes a bug in `metadataOnly()` where it was sometimes returning
+the LogLayer instance instead of nothing. It should now return nothing.
+
+For `loglayer`, plugins, and transports:
+
+- Adds a new TypeScript type called `LogLevelType`, which is a union of the `LogLevel` enum and its string representation.
+  - This means you can use either `LogLevel.info` or just `info` (string)
+  - All usages of `LogLevel` as a parameter type have been replaced with `LogLevelType`.
+
 ## Apr 12, 2025
 
 Adds support for the [llms.txt](https://llmstxt.org/) format.